+++ src/bot/botrulez/short_help.rs
@@ -2,12 +2,8 @@ use async_trait::async_trait;
 use clap::Parser;
 
 use crate::api::Message;
-use crate::bot::command::{ClapCommand, Context};
+use crate::bot::command::{ClapCommand, Command, Context};
 use crate::conn;
-
-/// Show short bot help.
-#[derive(Parser)]
-pub struct Args {}
 
 pub struct ShortHelp(pub String);
 
@@ -16,6 +12,29 @@ impl ShortHelp {
         Self(text.to_string())
     }
 }
+
+#[async_trait]
+impl<B, E> Command<B, E> for ShortHelp
+where
+    E: From<conn::Error>,
+{
+    async fn execute(
+        &self,
+        arg: &str,
+        msg: &Message,
+        ctx: &Context,
+        _bot: &mut B,
+    ) -> Result<(), E> {
+        if arg.trim().is_empty() {
+            ctx.reply(msg.id, &self.0).await?;
+        }
+        Ok(())
+    }
+}
+
+/// Show short bot help.
+#[derive(Parser)]
+pub struct Args {}
 
 #[async_trait]
 impl<B, E> ClapCommand<B, E> for ShortHelp