+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Procedure when bumping the version number:
 
 ### Added
 - `bot::botrulez::Ping` now implements `bot::command::Command`
+- `bot::botrulez::ShortHelp` now implements `bot::command::Command`
 
 ### Changed
 - Instances log to target `euphoxide::live::<name>`